+++ src/bittytax/conv/parsers/okx.py
@@ -198,7 +198,7 @@ def parse_okx_funding(
         )
     elif row_dict["Type"] == "Staking Yield":
         data_row.t_record = TransactionOutRecord(
-            TrType.STAKING,
+            TrType.STAKING_REWARD,
             data_row.timestamp,
             buy_quantity=Decimal(row_dict["Amount"]),
             buy_asset=row_dict["Symbol"],